hexa topo error + partitioned hexa errors for big data + new example just to check
