--- minisom2.py
@@ -1,818 +1,0 @@
-from math import sqrt
-
-from numpy import (array, unravel_index, nditer, linalg, random, subtract, max,
-                   power, exp, pi, zeros, ones, arange, outer, meshgrid, dot,
-                   logical_and, mean, std, cov, argsort, linspace, transpose,
-                   einsum, prod, nan, sqrt, hstack, diff, argmin, multiply)
-from numpy import sum as npsum
-from numpy.linalg import norm
-from collections import defaultdict, Counter
-from warnings import warn
-from sys import stdout
-from time import time
-from datetime import timedelta
-import pickle
-import os
-
-# for unit tests
-from numpy.testing import assert_almost_equal, assert_array_almost_equal
-from numpy.testing import assert_array_equal
-import unittest
-
-"""
-    Minimalistic implementation of the Self Organizing Maps (SOM).
-"""
-
-
-def _build_iteration_indexes(data_len, num_iterations,
-                             verbose=False, random_generator=None):
-    """Returns an iterable with the indexes of the samples
-    to pick at each iteration of the training.
-
-    If random_generator is not None, it must be an instalce
-    of numpy.random.RandomState and it will be used
-    to randomize the order of the samples."""
-    iterations = arange(num_iterations) % data_len
-    if random_generator:
-        random_generator.shuffle(iterations)
-    if verbose:
-        return _wrap_index__in_verbose(iterations)
-    else:
-        return iterations
-
-
-def _wrap_index__in_verbose(iterations):
-    """Yields the values in iterations printing the status on the stdout."""
-    m = len(iterations)
-    digits = len(str(m))
-    progress = '\r [ {s:{d}} / {m} ] {s:3.0f}% - ? it/s'
-    progress = progress.format(m=m, d=digits, s=0)
-    stdout.write(progress)
-    beginning = time()
-    stdout.write(progress)
-    for i, it in enumerate(iterations):
-        yield it
-        sec_left = ((m-i+1) * (time() - beginning)) / (i+1)
-        time_left = str(timedelta(seconds=sec_left))[:7]
-        progress = '\r [ {i:{d}} / {m} ]'.format(i=i+1, d=digits, m=m)
-        progress += ' {p:3.0f}%'.format(p=100*(i+1)/m)
-        progress += ' - {time_left} left '.format(time_left=time_left)
-        stdout.write(progress)
-
-
-def fast_norm(x):
-    """Returns norm-2 of a 1-D numpy array.
-
-    * faster than linalg.norm in case of 1-D arrays (numpy 1.9.2rc1).
-    """
-    return sqrt(dot(x, x.T))
-
-
-def asymptotic_decay(learning_rate, t, max_iter):
-    """Decay function of the learning process.
-    Parameters
-    ----------
-    learning_rate : float
-        current learning rate.
-
-    t : int
-        current iteration.
-
-    max_iter : int
-        maximum number of iterations for the training.
-    """
-    return learning_rate / (1+t/(max_iter/2))
-
-def get_coordinates_from_index(n, x, y):
-    if n<0:
-        return(-1,-1)
-    return(n%y, int(n/y))
-
-def get_index_from_coordinates(c, x, y):
-    if c[0]<0 or c[1]<0 or c[0]>y or c[1]>x:
-        return(-1)
-    return(c[1]*y + c[0])
-
-def get_neighbors_from_index(n, x, y):
-    c = get_coordinates_from_index(n, x, y)
-    offset_x = -1 if c[1]%2==0 else 1
-    neighbors_c =[(c[0]-1,c[1]), (c[0]+1,c[1]), (c[0],c[1]-1), (c[0]+offset_x,c[1]-1), (c[0],c[1]+1), 
-                (c[0]+offset_x,c[1]+1)] 
-    neighbors_c = list(filter(lambda c: (c[0]>=0) and (c[1]>=0) and (c[1]<x) and (c[0]<y), neighbors_c))
-    neighbors = list(map(lambda c: get_index_from_coordinates(c, x, y), neighbors_c))
-    return list(set(neighbors))
-
-
-class MiniSom(object):
-    def __init__(self, x, y, input_len, sigma=1.0, learning_rate=0.5,
-                 decay_function=asymptotic_decay,
-                 neighborhood_function='gaussian', topology='rectangular',
-                 activation_distance='euclidean', random_seed=None):
-        """Initializes a Self Organizing Maps.
-
-        A rule of thumb to set the size of the grid for a dimensionality
-        reduction task is that it should contain 5*sqrt(N) neurons
-        where N is the number of samples in the dataset to analyze.
-
-        E.g. if your dataset has 150 samples, 5*sqrt(150) = 61.23
-        hence a map 8-by-8 should perform well.
-
-        Parameters
-        ----------
-        x : int
-            x dimension of the SOM.
-
-        y : int
-            y dimension of the SOM.
-
-        input_len : int
-            Number of the elements of the vectors in input.
-
-        sigma : float, optional (default=1.0)
-            Spread of the neighborhood function, needs to be adequate
-            to the dimensions of the map.
-            (at the iteration t we have sigma(t) = sigma / (1 + t/T)
-            where T is #num_iteration/2)
-        learning_rate : initial learning rate
-            (at the iteration t we have
-            learning_rate(t) = learning_rate / (1 + t/T)
-            where T is #num_iteration/2)
-
-        decay_function : function (default=None)
-            Function that reduces learning_rate and sigma at each iteration
-            the default function is:
-                        learning_rate / (1+t/(max_iterarations/2))
-
-            A custom decay function will need to to take in input
-            three parameters in the following order:
-
-            1. learning rate
-            2. current iteration
-            3. maximum number of iterations allowed
-
-
-            Note that if a lambda function is used to define the decay
-            MiniSom will not be pickable anymore.
-
-        neighborhood_function : string, optional (default='gaussian')
-            Function that weights the neighborhood of a position in the map.
-            Possible values: 'gaussian', 'mexican_hat', 'bubble', 'triangle'
-
-        topology : string, optional (default='rectangular')
-            Topology of the map.
-            Possible values: 'rectangular', 'hexagonal'
-
-        activation_distance : string, optional (default='euclidean')
-            Distance used to activate the map.
-            Possible values: 'euclidean', 'cosine', 'manhattan', 'chebyshev'
-
-        random_seed : int, optional (default=None)
-            Random seed to use.
-        """
-        if sigma >= x or sigma >= y:
-            warn('Warning: sigma is too high for the dimension of the map.')
-
-        self._random_generator = random.RandomState(random_seed)
-
-        self._learning_rate = learning_rate
-        self._sigma = sigma
-        self._input_len = input_len
-        # random initialization
-        self._weights = self._random_generator.rand(x, y, input_len)*2-1
-        self._weights /= linalg.norm(self._weights, axis=-1, keepdims=True)
-
-        self._activation_map = zeros((x, y))
-        self._neigx = arange(x)
-        self._neigy = arange(y)  # used to evaluate the neighborhood function
-
-        if topology not in ['hexagonal', 'rectangular']:
-            msg = '%s not supported only hexagonal and rectangular available'
-            raise ValueError(msg % topology)
-        self.topology = topology
-        self._xx, self._yy = meshgrid(self._neigx, self._neigy)
-        self._xx = self._xx.astype(float)
-        self._yy = self._yy.astype(float)
-        if topology == 'hexagonal':
-            self._xx[::-2] -= 0.5
-            if neighborhood_function in ['triangle']:
-                warn('triangle neighborhood function does not ' +
-                     'take in account hexagonal topology')
-
-        self._decay_function = decay_function
-
-        neig_functions = {'gaussian': self._gaussian,
-                          'mexican_hat': self._mexican_hat,
-                          'bubble': self._bubble,
-                          'triangle': self._triangle}
-
-        if neighborhood_function not in neig_functions:
-            msg = '%s not supported. Functions available: %s'
-            raise ValueError(msg % (neighborhood_function,
-                                    ', '.join(neig_functions.keys())))
-
-        if neighborhood_function in ['triangle',
-                                     'bubble'] and (divmod(sigma, 1)[1] != 0
-                                                    or sigma < 1):
-            warn('sigma should be an integer >=1 when triangle or bubble' +
-                 'are used as neighborhood function')
-
-        self.neighborhood = neig_functions[neighborhood_function]
-
-        distance_functions = {'euclidean': self._euclidean_distance,
-                              'cosine': self._cosine_distance,
-                              'manhattan': self._manhattan_distance,
-                              'chebyshev': self._chebyshev_distance}
-
-        if activation_distance not in distance_functions:
-            msg = '%s not supported. Distances available: %s'
-            raise ValueError(msg % (activation_distance,
-                                    ', '.join(distance_functions.keys())))
-
-        self._activation_distance = distance_functions[activation_distance]
-
-    def get_weights(self):
-        """Returns the weights of the neural network."""
-        return self._weights
-
-    def get_euclidean_coordinates(self):
-        """Returns the position of the neurons on an euclidean
-        plane that reflects the chosen topology in two meshgrids xx and yy.
-        Neuron with map coordinates (1, 4) has coordinate (xx[1, 4], yy[1, 4])
-        in the euclidean plane.
-
-        Only useful if the topology chosen is not rectangular.
-        """
-        return self._xx.T, self._yy.T
-
-    def convert_map_to_euclidean(self, xy):
-        """Converts map coordinates into euclidean coordinates
-        that reflects the chosen topology.
-
-        Only useful if the topology chosen is not rectangular.
-        """
-        return self._xx.T[xy], self._yy.T[xy]
-
-    def _activate(self, x):
-        """Updates matrix activation_map, in this matrix
-           the element i,j is the response of the neuron i,j to x."""
-        self._activation_map = self._activation_distance(x, self._weights)
-
-    def activate(self, x):
-        """Returns the activation map to x."""
-        self._activate(x)
-        return self._activation_map
-
-    def _gaussian(self, c, sigma):
-        """Returns a Gaussian centered in c."""
-        d = 2*pi*sigma*sigma
-        ax = exp(-power(self._xx-self._xx.T[c], 2)/d)
-        ay = exp(-power(self._yy-self._yy.T[c], 2)/d)
-        return (ax * ay).T  # the external product gives a matrix
-
-    def _mexican_hat(self, c, sigma):
-        """Mexican hat centered in c."""
-        p = power(self._xx-self._xx.T[c], 2) + power(self._yy-self._yy.T[c], 2)
-        d = 2*pi*sigma*sigma
-        return (exp(-p/d)*(1-2/d*p)).T
-
-    def _bubble(self, c, sigma):
-        """Constant function centered in c with spread sigma.
-        sigma should be an odd value.
-        """
-        ax = logical_and(self._neigx > c[0]-sigma,
-                         self._neigx < c[0]+sigma)
-        ay = logical_and(self._neigy > c[1]-sigma,
-                         self._neigy < c[1]+sigma)
-        return outer(ax, ay)*1.
-
-    def _triangle(self, c, sigma):
-        """Triangular function centered in c with spread sigma."""
-        triangle_x = (-abs(c[0] - self._neigx)) + sigma
-        triangle_y = (-abs(c[1] - self._neigy)) + sigma
-        triangle_x[triangle_x < 0] = 0.
-        triangle_y[triangle_y < 0] = 0.
-        return outer(triangle_x, triangle_y)
-
-    def _cosine_distance(self, x, w):
-        num = (w * x).sum(axis=2)
-        denum = multiply(linalg.norm(w, axis=2), linalg.norm(x))
-        return 1 - num / (denum+1e-8)
-
-    def _euclidean_distance(self, x, w):
-        return linalg.norm(subtract(x, w), axis=-1)
-
-    def _manhattan_distance(self, x, w):
-        return linalg.norm(subtract(x, w), ord=1, axis=-1)
-
-    def _chebyshev_distance(self, x, w):
-        return max(subtract(x, w), axis=-1)
-
-    def _check_iteration_number(self, num_iteration):
-        if num_iteration < 1:
-            raise ValueError('num_iteration must be > 1')
-
-    def _check_input_len(self, data):
-        """Checks that the data in input is of the correct shape."""
-        data_len = len(data[0])
-        if self._input_len != data_len:
-            msg = 'Received %d features, expected %d.' % (data_len,
-                                                          self._input_len)
-            raise ValueError(msg)
-
-    def winner(self, x):
-        """Computes the coordinates of the winning neuron for the sample x."""
-        self._activate(x)
-        return unravel_index(self._activation_map.argmin(),
-                             self._activation_map.shape)
-
-    def update(self, x, win, t, max_iteration):
-        """Updates the weights of the neurons.
-
-        Parameters
-        ----------
-        x : np.array
-            Current pattern to learn.
-        win : tuple
-            Position of the winning neuron for x (array or tuple).
-        t : int
-            Iteration index
-        max_iteration : int
-            Maximum number of training itarations.
-        """
-        eta = self._decay_function(self._learning_rate, t, max_iteration)
-        # sigma and learning rate decrease with the same rule
-        sig = self._decay_function(self._sigma, t, max_iteration)
-        # improves the performances
-        g = self.neighborhood(win, sig)*eta
-        # w_new = eta * neighborhood_function * (x-w)
-        self._weights += einsum('ij, ijk->ijk', g, x-self._weights)
-
-    def quantization(self, data):
-        """Assigns a code book (weights vector of the winning neuron)
-        to each sample in data."""
-        self._check_input_len(data)
-        winners_coords = argmin(self._distance_from_weights(data), axis=1)
-        return self._weights[unravel_index(winners_coords,
-                                           self._weights.shape[:2])]
-
-    def random_weights_init(self, data):
-        """Initializes the weights of the SOM
-        picking random samples from data."""
-        self._check_input_len(data)
-        it = nditer(self._activation_map, flags=['multi_index'])
-        while not it.finished:
-            rand_i = self._random_generator.randint(len(data))
-            self._weights[it.multi_index] = data[rand_i]
-            it.iternext()
-
-    def pca_weights_init(self, data):
-        """Initializes the weights to span the first two principal components.
-
-        This initialization doesn't depend on random processes and
-        makes the training process converge faster.
-
-        It is strongly reccomended to normalize the data before initializing
-        the weights and use the same normalization for the training data.
-        """
-        if self._input_len == 1:
-            msg = 'The data needs at least 2 features for pca initialization'
-            raise ValueError(msg)
-        self._check_input_len(data)
-        if len(self._neigx) == 1 or len(self._neigy) == 1:
-            msg = 'PCA initialization inappropriate:' + \
-                  'One of the dimensions of the map is 1.'
-            warn(msg)
-        pc_length, pc = linalg.eig(cov(transpose(data)))
-        pc_order = argsort(-pc_length)
-        for i, c1 in enumerate(linspace(-1, 1, len(self._neigx))):
-            for j, c2 in enumerate(linspace(-1, 1, len(self._neigy))):
-                self._weights[i, j] = c1*pc[pc_order[0]] + c2*pc[pc_order[1]]
-
-    def train(self, data, num_iteration, random_order=False, verbose=False):
-        """Trains the SOM.
-
-        Parameters
-        ----------
-        data : np.array or list
-            Data matrix.
-
-        num_iteration : int
-            Maximum number of iterations (one iteration per sample).
-        random_order : bool (default=False)
-            If True, samples are picked in random order.
-            Otherwise the samples are picked sequentially.
-
-        verbose : bool (default=False)
-            If True the status of the training
-            will be printed at each iteration.
-        """
-        self._check_iteration_number(num_iteration)
-        self._check_input_len(data)
-        random_generator = None
-        if random_order:
-            random_generator = self._random_generator
-        iterations = _build_iteration_indexes(len(data), num_iteration,
-                                              verbose, random_generator)
-        for t, iteration in enumerate(iterations):
-            self.update(data[iteration], self.winner(data[iteration]),
-                        t, num_iteration)
-        if verbose:
-            print('\n quantization error:', self.quantization_error(data))
-
-    def train_random(self, data, num_iteration, verbose=False):
-        """Trains the SOM picking samples at random from data.
-
-        Parameters
-        ----------
-        data : np.array or list
-            Data matrix.
-
-        num_iteration : int
-            Maximum number of iterations (one iteration per sample).
-
-        verbose : bool (default=False)
-            If True the status of the training
-            will be printed at each iteration.
-        """
-        self.train(data, num_iteration, random_order=True, verbose=verbose)
-
-    def train_batch(self, data, num_iteration, verbose=False):
-        """Trains the SOM using all the vectors in data sequentially.
-
-        Parameters
-        ----------
-        data : np.array or list
-            Data matrix.
-
-        num_iteration : int
-            Maximum number of iterations (one iteration per sample).
-
-        verbose : bool (default=False)
-            If True the status of the training
-            will be printed at each iteration.
-        """
-        self.train(data, num_iteration, random_order=False, verbose=verbose)
-
-    def distance_map(self):
-        """Returns the distance map of the weights.
-        Each cell is the normalised sum of the distances between
-        a neuron and its neighbours. Note that this method uses
-        the euclidean distance."""
-        um = zeros((self._weights.shape[0],
-                    self._weights.shape[1],
-                    8))  # 2 spots more for hexagonal topology
-
-        ii = [[0, -1, -1, -1, 0, 1, 1, 1]]*2
-        jj = [[-1, -1, 0, 1, 1, 1, 0, -1]]*2
-
-        if self.topology == 'hexagonal':
-            ii = [[1, 1, 1, 0, -1, 0], [0, 1, 0, -1, -1, -1]]
-            jj = [[1, 0, -1, -1, 0, 1], [1, 0, -1, -1, 0, 1]]
-
-        for x in range(self._weights.shape[0]):
-            for y in range(self._weights.shape[1]):
-                w_2 = self._weights[x, y]
-                e = y % 2 == 0   # only used on hexagonal topology
-                for k, (i, j) in enumerate(zip(ii[e], jj[e])):
-                    if (x+i >= 0 and x+i < self._weights.shape[0] and
-                            y+j >= 0 and y+j < self._weights.shape[1]):
-                        w_1 = self._weights[x+i, y+j]
-                        um[x, y, k] = fast_norm(w_2-w_1)
-
-        um = um.sum(axis=2)
-        return um/um.max()
-
-    def activation_response(self, data):
-        """
-            Returns a matrix where the element i,j is the number of times
-            that the neuron i,j have been winner.
-        """
-        self._check_input_len(data)
-        a = zeros((self._weights.shape[0], self._weights.shape[1]))
-        for x in data:
-            a[self.winner(x)] += 1
-        return a
-
-    def _distance_from_weights(self, data):
-        """Returns a matrix d where d[i,j] is the euclidean distance between
-        data[i] and the j-th weight.
-        """
-        input_data = array(data)
-        weights_flat = self._weights.reshape(-1, self._weights.shape[2])
-        input_data_sq = power(input_data, 2).sum(axis=1, keepdims=True)
-        weights_flat_sq = power(weights_flat, 2).sum(axis=1, keepdims=True)
-        cross_term = dot(input_data, weights_flat.T)
-        return sqrt(-2 * cross_term + input_data_sq + weights_flat_sq.T)
-
-    def quantization_error(self, data):
-        """Returns the quantization error computed as the average
-        distance between each input sample and its best matching unit."""
-        self._check_input_len(data)
-        return norm(data-self.quantization(data), axis=1).mean()
-
-    def topographic_error(self, data):
-        """Returns the topographic error computed by finding
-        the best-matching and second-best-matching neuron in the map
-        for each input and then evaluating the positions.
-
-        A sample for which these two nodes are not adjacent counts as
-        an error. The topographic error is given by the
-        the total number of errors divided by the total of samples.
-
-        If the topographic error is 0, no error occurred.
-        If 1, the topology was not preserved for any of the samples."""
-        self._check_input_len(data)
-        if self.topology == 'hexagonal':
-            msg = 'Topographic error not implemented for hexagonal topology.'
-            raise NotImplementedError(msg)
-        total_neurons = prod(self._activation_map.shape)
-        if total_neurons == 1:
-            warn('The topographic error is not defined for a 1-by-1 map.')
-            return nan
-
-        t = 1.42
-        # b2mu: best 2 matching units
-        b2mu_inds = argsort(self._distance_from_weights(data), axis=1)[:, :2]
-        b2my_xy = unravel_index(b2mu_inds, self._weights.shape[:2])
-        b2mu_x, b2mu_y = b2my_xy[0], b2my_xy[1]
-        dxdy = hstack([diff(b2mu_x), diff(b2mu_y)])
-        distance = norm(dxdy, axis=1)
-        return (distance > t).mean()
-
-    def win_map(self, data, return_indices=False):
-        """Returns a dictionary wm where wm[(i,j)] is a list with:
-        - all the patterns that have been mapped to the position (i,j),
-          if return_indices=False (default)
-        - all indices of the elements that have been mapped to the
-          position (i,j) if return_indices=True"""
-        self._check_input_len(data)
-        winmap = defaultdict(list)
-        for i, x in enumerate(data):
-            winmap[self.winner(x)].append(i if return_indices else x)
-        return winmap
-
-    def labels_map(self, data, labels):
-        """Returns a dictionary wm where wm[(i,j)] is a dictionary
-        that contains the number of samples from a given label
-        that have been mapped in position i,j.
-
-        Parameters
-        ----------
-        data : np.array or list
-            Data matrix.
-
-        label : np.array or list
-            Labels for each sample in data.
-        """
-        self._check_input_len(data)
-        if not len(data) == len(labels):
-            raise ValueError('data and labels must have the same length.')
-        winmap = defaultdict(list)
-        for x, l in zip(data, labels):
-            winmap[self.winner(x)].append(l)
-        for position in winmap:
-            winmap[position] = Counter(winmap[position])
-        return winmap
-
-    def partitioned_quant_error(self,data,chunks):
-        n=len(data)
-        n_chunks = list(range(0,n,chunks))
-        n_chunks.append(n)
-        errorq=0
-        for j in range(len(n_chunks)-1):
-            interval = n_chunks[j+1]-n_chunks[j]
-            errorq = errorq+self.quantization_error(data[n_chunks[j]:n_chunks[j+1]])*interval/n
-        return errorq
-
-    def calculate_hexa_topographical_error(self, data):
-        bmus_1st_and_2nd = argsort(self._distance_from_weights(data), axis=1)[:, :2]
-        n_rows = self._weights.shape[0]
-        n_columns = self._weights.shape[1]
-        neighbors=list(map(lambda t:get_neighbors_from_index(t, n_rows, n_columns), bmus_1st_and_2nd[:,0]))
-
-        e_t=1-mean([second in neighs for (second,neighs) in zip(bmus_1st_and_2nd[:,1], neighbors)])
-        return(e_t)
-
-    def partitioned_topo_error(self,data,chunks):
-        n=len(data)
-        n_chunks = list(range(0,n,chunks))
-        n_chunks.append(n)
-        errort=0
-        for j in range(len(n_chunks)-1):
-            interval = n_chunks[j+1]-n_chunks[j]
-            errort = errort+self.calculate_hexa_topographical_error(data[n_chunks[j]:n_chunks[j+1]])*interval/n
-        return errort
-
-
-
-
-class TestMinisom(unittest.TestCase):
-    def setUp(self):
-        self.som = MiniSom(5, 5, 1)
-        for i in range(5):
-            for j in range(5):
-                # checking weights normalization
-                assert_almost_equal(1.0, linalg.norm(self.som._weights[i, j]))
-        self.som._weights = zeros((5, 5, 1))  # fake weights
-        self.som._weights[2, 3] = 5.0
-        self.som._weights[1, 1] = 2.0
-
-    def test_decay_function(self):
-        assert self.som._decay_function(1., 2., 3.) == 1./(1.+2./(3./2))
-
-    def test_fast_norm(self):
-        assert fast_norm(array([1, 3])) == sqrt(1+9)
-
-    def test_euclidean_distance(self):
-        x = zeros((1, 2))
-        w = ones((2, 2, 2))
-        d = self.som._euclidean_distance(x, w)
-        assert_array_almost_equal(d, [[1.41421356, 1.41421356],
-                                      [1.41421356, 1.41421356]])
-
-    def test_cosine_distance(self):
-        x = zeros((1, 2))
-        w = ones((2, 2, 2))
-        d = self.som._cosine_distance(x, w)
-        assert_array_almost_equal(d, [[1., 1.],
-                                      [1., 1.]])
-
-    def test_manhattan_distance(self):
-        x = zeros((1, 2))
-        w = ones((2, 2, 2))
-        d = self.som._manhattan_distance(x, w)
-        assert_array_almost_equal(d, [[2., 2.],
-                                      [2., 2.]])
-
-    def test_chebyshev_distance(self):
-        x = array([1, 3])
-        w = ones((2, 2, 2))
-        d = self.som._chebyshev_distance(x, w)
-        assert_array_almost_equal(d, [[2., 2.],
-                                      [2., 2.]])
-
-    def test_check_input_len(self):
-        with self.assertRaises(ValueError):
-            self.som.train_batch([[1, 2]], 1)
-
-        with self.assertRaises(ValueError):
-            self.som.random_weights_init(array([[1, 2]]))
-
-        with self.assertRaises(ValueError):
-            self.som._check_input_len(array([[1, 2]]))
-
-        self.som._check_input_len(array([[1]]))
-        self.som._check_input_len([[1]])
-
-    def test_unavailable_neigh_function(self):
-        with self.assertRaises(ValueError):
-            MiniSom(5, 5, 1, neighborhood_function='boooom')
-
-    def test_unavailable_distance_function(self):
-        with self.assertRaises(ValueError):
-            MiniSom(5, 5, 1, activation_distance='ridethewave')
-
-    def test_gaussian(self):
-        bell = self.som._gaussian((2, 2), 1)
-        assert bell.max() == 1.0
-        assert bell.argmax() == 12  # unravel(12) = (2,2)
-
-    def test_mexican_hat(self):
-        bell = self.som._mexican_hat((2, 2), 1)
-        assert bell.max() == 1.0
-        assert bell.argmax() == 12  # unravel(12) = (2,2)
-
-    def test_bubble(self):
-        bubble = self.som._bubble((2, 2), 1)
-        assert bubble[2, 2] == 1
-        assert sum(sum(bubble)) == 1
-
-    def test_triangle(self):
-        bubble = self.som._triangle((2, 2), 1)
-        assert bubble[2, 2] == 1
-        assert sum(sum(bubble)) == 1
-
-    def test_win_map(self):
-        winners = self.som.win_map([[5.0], [2.0]])
-        assert winners[(2, 3)][0] == [5.0]
-        assert winners[(1, 1)][0] == [2.0]
-
-    def test_win_map_indices(self):
-        winners = self.som.win_map([[5.0], [2.0]], return_indices=True)
-        assert winners[(2, 3)] == [0]
-        assert winners[(1, 1)] == [1]
-
-    def test_labels_map(self):
-        labels_map = self.som.labels_map([[5.0], [2.0]], ['a', 'b'])
-        assert labels_map[(2, 3)]['a'] == 1
-        assert labels_map[(1, 1)]['b'] == 1
-        with self.assertRaises(ValueError):
-            self.som.labels_map([[5.0]], ['a', 'b'])
-
-    def test_activation_reponse(self):
-        response = self.som.activation_response([[5.0], [2.0]])
-        assert response[2, 3] == 1
-        assert response[1, 1] == 1
-
-    def test_activate(self):
-        assert self.som.activate(5.0).argmin() == 13.0  # unravel(13) = (2,3)
-
-    def test_distance_from_weights(self):
-        data = arange(-5, 5).reshape(-1, 1)
-        weights = self.som._weights.reshape(-1, self.som._weights.shape[2])
-        distances = self.som._distance_from_weights(data)
-        for i in range(len(data)):
-            for j in range(len(weights)):
-                assert(distances[i][j] == norm(data[i] - weights[j]))
-
-    def test_quantization_error(self):
-        assert self.som.quantization_error([[5], [2]]) == 0.0
-        assert self.som.quantization_error([[4], [1]]) == 1.0
-
-    def test_topographic_error(self):
-        # 5 will have bmu_1 in (2,3) and bmu_2 in (2, 4)
-        # which are in the same neighborhood
-        self.som._weights[2, 4] = 6.0
-        # 15 will have bmu_1 in (4, 4) and bmu_2 in (0, 0)
-        # which are not in the same neighborhood
-        self.som._weights[4, 4] = 15.0
-        self.som._weights[0, 0] = 14.
-        assert self.som.topographic_error([[5]]) == 0.0
-        assert self.som.topographic_error([[15]]) == 1.0
-
-        self.som.topology = 'hexagonal'
-        with self.assertRaises(NotImplementedError):
-            assert self.som.topographic_error([[5]]) == 0.0
-        self.som.topology = 'rectangular'
-
-    def test_quantization(self):
-        q = self.som.quantization(array([[4], [2]]))
-        assert q[0] == 5.0
-        assert q[1] == 2.0
-
-    def test_random_seed(self):
-        som1 = MiniSom(5, 5, 2, sigma=1.0, learning_rate=0.5, random_seed=1)
-        som2 = MiniSom(5, 5, 2, sigma=1.0, learning_rate=0.5, random_seed=1)
-        # same initialization
-        assert_array_almost_equal(som1._weights, som2._weights)
-        data = random.rand(100, 2)
-        som1 = MiniSom(5, 5, 2, sigma=1.0, learning_rate=0.5, random_seed=1)
-        som1.train_random(data, 10)
-        som2 = MiniSom(5, 5, 2, sigma=1.0, learning_rate=0.5, random_seed=1)
-        som2.train_random(data, 10)
-        # same state after training
-        assert_array_almost_equal(som1._weights, som2._weights)
-
-    def test_train_batch(self):
-        som = MiniSom(5, 5, 2, sigma=1.0, learning_rate=0.5, random_seed=1)
-        data = array([[4, 2], [3, 1]])
-        q1 = som.quantization_error(data)
-        som.train(data, 10)
-        assert q1 > som.quantization_error(data)
-
-        data = array([[1, 5], [6, 7]])
-        q1 = som.quantization_error(data)
-        som.train_batch(data, 10, verbose=True)
-        assert q1 > som.quantization_error(data)
-
-    def test_train_random(self):
-        som = MiniSom(5, 5, 2, sigma=1.0, learning_rate=0.5, random_seed=1)
-        data = array([[4, 2], [3, 1]])
-        q1 = som.quantization_error(data)
-        som.train(data, 10, random_order=True)
-        assert q1 > som.quantization_error(data)
-
-        data = array([[1, 5], [6, 7]])
-        q1 = som.quantization_error(data)
-        som.train_random(data, 10, verbose=True)
-        assert q1 > som.quantization_error(data)
-
-    def test_random_weights_init(self):
-        som = MiniSom(2, 2, 2, random_seed=1)
-        som.random_weights_init(array([[1.0, .0]]))
-        for w in som._weights:
-            assert_array_equal(w[0], array([1.0, .0]))
-
-    def test_pca_weights_init(self):
-        som = MiniSom(2, 2, 2)
-        som.pca_weights_init(array([[1.,  0.], [0., 1.], [1., 0.], [0., 1.]]))
-        expected = array([[[0., -1.41421356], [-1.41421356, 0.]],
-                          [[1.41421356, 0.], [0., 1.41421356]]])
-        assert_array_almost_equal(som._weights, expected)
-
-    def test_distance_map(self):
-        som = MiniSom(2, 2, 2, random_seed=1)
-        som._weights = array([[[1.,  0.], [0., 1.]], [[1., 0.], [0., 1.]]])
-        assert_array_equal(som.distance_map(), array([[1., 1.], [1., 1.]]))
-
-        som = MiniSom(2, 2, 2, topology='hexagonal', random_seed=1)
-        som._weights = array([[[1.,  0.], [0., 1.]], [[1., 0.], [0., 1.]]])
-        assert_array_equal(som.distance_map(), array([[.5, 1.], [1., .5]]))
-
-    def test_pickling(self):
-        with open('som.p', 'wb') as outfile:
-            pickle.dump(self.som, outfile)
-        with open('som.p', 'rb') as infile:
-            pickle.load(infile)
-        os.remove('som.p')
-
